requirements.txt
twitchio>=2.6.0,<3
requests>=2.31.0
python-dotenv>=1.0.0
streamlink>=7.0.0
aiohttp>=3.9.0
numpy>=1.26.0
openai-whisper>=20231117
